Handle editing a new file.

+++ safe_edit.py
@@ -37,10 +37,11 @@ def safe_edit(file_name):
     tf_name = None
     tmp_file = None
     try:
-        try:
-            in_file = open(file_name, mode='r', encoding='utf-8')
-        except TypeError:
-            in_file = open(file_name, mode='r')
+        if os.path.isfile(file_name):
+            try:
+                in_file = open(file_name, mode='r', encoding='utf-8')
+            except TypeError:
+                in_file = open(file_name, mode='r')
         try:
             # noinspection PyArgumentList
             tmp_file = NamedTemporaryFile(mode='w', delete=False, encoding='utf-8')
@@ -77,9 +78,11 @@ def safe_edit(file_name):
             # Note, shutil.move will safely move even across file systems
 
             # backup source file
-            shutil.move(file_name, backup_name)
+            if os.path.isfile(file_name):
+                shutil.move(file_name, backup_name)
 
             # put new file in place
+            # noinspection PyTypeChecker
             shutil.move(tf_name, file_name)
